Refactor commit function into App for cleaner code organization
